Refactor: rename `ScanningEncoder` to `UnifiedVocabEncoder`, update struct, imports, and references across modules for consistency.
