+++ crates/bpetok/src/decoder/corpus_decoder.rs
@@ -231,7 +231,7 @@ impl<T: TokenType> MaterializationMap<T> {
 mod tests {
     use super::*;
     use crate::tokenizer::TokenEncoder;
-    use crate::tokenizer::scanning_encoder::ScanningEncoder;
+    use crate::tokenizer::unified_encoder::ScanningEncoder;
     use crate::types::{check_is_send, check_is_sync};
     use crate::vocab::data::unified::UnifiedTokenVocab;
     use crate::vocab::training::trainer::{BPETokenVocabTrainer, TrainResults};
